convert unit_price to decimal data type

+++ db/migrate/20200317043155_create_item.rb
@@ -3,7 +3,7 @@ class CreateItem < ActiveRecord::Migration[6.0]
     create_table :items do |t|
       t.string :name
       t.string :description
-      t.float :unit_price
+      t.decimal :unit_price
 			t.references :merchant, null: false, foreign_key: true
 
 			t.timestamps
